# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Universal GPS Navigation & Delivery Optimization API
  version: 0.1.0
  description: Core REST endpoints for orders, optimization, routes, driver events, PoD, and webhooks.
servers:
  - url: http://localhost:8081
    description: Local Docker (compose)
  - url: http://localhost:8082
    description: Local Docker (compose alt)
  - url: http://localhost:8080
    description: Local direct (make run)
  - url: https://api.example.com
    description: Production (example)
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
  /v1/orders:
    post:
      tags: [Orders]
      summary: Bulk import orders
      parameters:
        - in: header
          name: Idempotency-Key
          schema: { type: string }
          required: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderImportRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderImportResponse' }
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderListResponse' }
  /v1/optimize:
    post:
      tags: [Optimization]
      summary: Plan or re-plan routes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OptimizeRequest' }
      responses:
        '200':
          description: Planned routes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OptimizeResponse' }
...
  /v1/driver-events:
    post:
      tags: [DriverEvents]
      summary: Ingest driver events batch
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DriverEventsRequest' }
      responses:
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/DriverEventsAck' } } } }
  /v1/pod:
    post:
      tags: [PoD]
      summary: Upload PoD metadata or request upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PoDRequest' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/PoDResponse' } } } }
  /v1/subscriptions:
    post:
      tags: [Subscriptions]
      summary: Create webhook subscription
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubscriptionRequest' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Subscription' } } } }
    get:
      tags: [Subscriptions, SubscriptionsAdmin]
      summary: List webhook subscriptions
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SubscriptionListResponse' } } } }
...
  /v1/admin/webhook-dlq:
    get:
      tags: [WebhooksAdmin]
      summary: List dead-lettered webhook deliveries
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
        - in: query
          name: eventType
          schema: { type: string }
        - in: query
          name: olderThanHours
          schema: { type: integer }
        - in: query
          name: responseCodeMin
          schema: { type: integer }
        - in: query
          name: responseCodeMax
          schema: { type: integer }
        - in: query
          name: errorQuery
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { type: object } }, nextCursor: { type: string } } } } } }
    post:
      tags: [WebhooksAdmin]
      summary: Bulk requeue DLQ entries by id list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items: { type: string }
      responses:
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/Ack' } } } }
    delete:
      tags: [WebhooksAdmin]
      summary: Bulk delete DLQ entries by ids or olderThanHours
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items: { type: string }
                olderThanHours: { type: integer }
      responses:
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/Ack' } } } }
...
  /v1/eta/stream:
    get:
      tags: [Routes]
      summary: ETA Server-Sent Events
      parameters:
        - in: query
          name: routeId
          required: true
          schema: { type: string }
      responses:
        '200': { description: SSE stream }
...
  /v1/geofences:
    get:
      tags: [Geofences]
      summary: List geofences
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GeofenceListResponse' } } } }
    post:
      tags: [Geofences]
      summary: Create a geofence
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GeofenceInput' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Geofence' } } } }
...
  /v1/media/presign:
    post:
      tags: [Media]
      summary: Request a presigned upload URL for PoD media
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PresignRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PresignResponse' } } } }
...
  /v1/admin/routes/stats:
    get:
      tags: [Metrics]
      summary: Route and leg aggregates for a plan date
      parameters:
        - in: query
          name: planDate
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RouteStats' } } } }
...
  /v1/admin/optimizer/config:
    get:
      tags: [Optimization]
      summary: Get tenant optimizer default configuration
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  config: { type: object }
    put:
      tags: [Optimization]
      summary: Update tenant optimizer default configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                config: { type: object }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Ack:
      type: object
      properties:
        accepted: { type: integer }
        rejected: { type: integer }
    OrderImportRequest:
      type: object
      required: [tenantId, orders]
      properties:
        tenantId: { type: string }
        orders:
          type: array
          items: { $ref: '#/components/schemas/OrderInput' }
    OrderInput:
      type: object
      required: [stops]
      properties:
        externalRef: { type: string }
        priority: { type: integer, default: 0 }
        attributes: { type: object, additionalProperties: true }
        stops:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/StopInput' }
    StopInput:
      type: object
      required: [type, location]
      properties:
        type: { type: string, enum: [pickup, delivery] }
        address: { type: string }
        location:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
        timeWindow:
          type: object
          properties:
            start: { type: string, format: date-time }
            end: { type: string, format: date-time }
        serviceTimeSec: { type: integer, default: 0 }
        requiredSkills:
          type: array
          items: { type: string }
    OrderImportResponse:
      type: object
      properties:
        importId: { type: string }
        created: { type: integer }
        skipped: { type: integer }
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Order' }
        nextCursor: { type: string, nullable: true }
    Order:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        externalRef: { type: string }
        priority: { type: integer }
        status: { type: string }
        stops:
          type: array
          items: { $ref: '#/components/schemas/Stop' }
    Stop:
      type: object
      properties:
        id: { type: string }
        orderId: { type: string }
        type: { type: string }
        address: { type: string }
        location:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
        timeWindow:
          type: object
          properties:
            start: { type: string, format: date-time }
            end: { type: string, format: date-time }
        serviceTimeSec: { type: integer }
        status: { type: string }
    OptimizeRequest:
      type: object
      required: [tenantId, planDate]
      properties:
        tenantId: { type: string }
        planDate: { type: string, format: date }
        algorithm:
          type: string
          enum: [greedy, alns]
          description: Strategy for planning; defaults to greedy if unspecified
        timeBudgetMs:
          type: integer
          description: Max wall-clock time for optimizer in milliseconds (ALNS)
          example: 500
        maxIterations:
          type: integer
          description: Max ALNS iterations; if set, acts as an upper bound in addition to timeBudgetMs
        vehiclePool:
          type: array
          items: { type: string }
        depots:
          type: array
          items: { type: string }
        includeOrders:
          type: array
          items: { type: string }
        constraints: { type: object, additionalProperties: true }
        objectives: { type: object, additionalProperties: { type: number } }
        reoptimize: { type: boolean, default: false }
        freeze:
          type: object
          properties:
            routes:
              type: array
              items: { type: string }
            upToLegId: { type: string }
    OptimizeResponse:
      type: object
      properties:
        batchId: { type: string }
        routes:
          type: array
          items: { $ref: '#/components/schemas/Route' }
    Route:
      type: object
      properties:
        id: { type: string }
        version: { type: integer }
        planDate: { type: string, format: date }
        status: { type: string }
        driverId: { type: string }
        vehicleId: { type: string }
        legs:
          type: array
          items: { $ref: '#/components/schemas/Leg' }
        costBreakdown: { type: object, additionalProperties: { type: number } }
        autoAdvance: { $ref: '#/components/schemas/AutoAdvancePolicy' }
        breaksCount: { type: integer, description: Number of planned breaks }
        totalBreakSec: { type: integer, description: Total planned break seconds }
    Leg:
      type: object
      properties:
        id: { type: string }
        seq: { type: integer }
        kind: { type: string, enum: [drive, break], default: drive }
        breakSec: { type: integer, description: Duration if kind=break }
        fromStopId: { type: string }
        toStopId: { type: string }
        distM: { type: integer }
        driveSec: { type: integer }
        etaArrival: { type: string, format: date-time }
        etaDeparture: { type: string, format: date-time }
        status: { type: string }
...
    DriverEventsRequest:
      type: object
      required: [tenantId, events]
      properties:
        tenantId: { type: string }
        events:
          type: array
          items: { $ref: '#/components/schemas/DriverEvent' }
    DriverEvent:
      type: object
      required: [type, ts]
      properties:
        type: { type: string, enum: [location, arrive, depart, exception, pod] }
        driverId: { type: string }
        routeId: { type: string }
        legId: { type: string }
        stopId: { type: string }
        ts: { type: string, format: date-time }
        payload: { type: object, additionalProperties: true }
    DriverEventsAck:
      type: object
      properties:
        accepted: { type: integer }
        rejected: { type: integer }
        advanced:
          type: array
          description: List of auto-advancements triggered by this batch
          items: { $ref: '#/components/schemas/AdvanceResult' }
    PoDRequest:
      type: object
      required: [tenantId, orderId, stopId, type]
      properties:
        tenantId: { type: string }
        orderId: { type: string }
        stopId: { type: string }
        type: { type: string, enum: [signature, photo, barcode] }
        media:
          type: object
          properties:
            uploadUrl: { type: string }
            sha256: { type: string }
        metadata: { type: object, additionalProperties: true }
    PoDResponse:
      type: object
      properties:
        podId: { type: string }
        status: { type: string }
    SubscriptionRequest:
      type: object
      required: [tenantId, url, events]
      properties:
        tenantId: { type: string }
        url: { type: string, format: uri }
        events:
          type: array
          items: { type: string }
        secret: { type: string }
    Subscription:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        url: { type: string, format: uri }
        events:
          type: array
          items: { type: string }
        secret: { type: string }
    SubscriptionListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Subscription' }
        nextCursor: { type: string, nullable: true }
...
    GeofenceInput:
      type: object
      properties:
        name: { type: string }
        type: { type: string, enum: [hub, customer, restricted] }
        radiusM: { type: integer }
        center:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
        rules: { type: object, additionalProperties: true }
    Geofence:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        name: { type: string }
        type: { type: string }
        radiusM: { type: integer }
        center: { $ref: '#/components/schemas/GeoPoint' }
        rules: { type: object, additionalProperties: true }
    GeofenceListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Geofence' }
        nextCursor: { type: string, nullable: true }
    PresignRequest:
      type: object
      required: [tenantId, fileName, contentType]
      properties:
        tenantId: { type: string }
        fileName: { type: string }
        contentType: { type: string }
        bytes: { type: integer }
        sha256: { type: string }
    PresignResponse:
      type: object
      properties:
        uploadUrl: { type: string }
        method: { type: string, example: PUT }
        headers: { type: object, additionalProperties: { type: string } }
        expireAt: { type: string, format: date-time }
...
    GeoPoint:
      type: object
      properties:
        lat: { type: number }
        lng: { type: number }
...
    RouteStats:
      type: object
      properties:
        routes: { type: integer }
        legs: { type: integer }
        totalDistM: { type: integer }
        totalDriveSec: { type: integer }
        avgLegsPerRoute: { type: number }
        breaks: { type: integer }
        breakSec: { type: integer }
    AutoAdvancePolicy:
      type: object
      description: Controls automatic progression to the next stop
      properties:
        enabled: { type: boolean, default: false }
        trigger:
          type: string
          enum: [pod_ack, depart, geofence_arrive]
          description: Which event triggers auto-advance
        minDwellSec: { type: integer, description: Minimum dwell time before advancing }
        requirePoD: { type: boolean, description: Require PoD before advance, default: true }
        gracePeriodSec: { type: integer, description: Wait before advancing after trigger }
        movingLock: { type: boolean, description: Prevent UI advances while vehicle is moving, default: true }
        hosMaxDriveSec: { type: integer, description: Max continuous drive seconds before requiring break }
...
    AdvanceResult:
      type: object
      properties:
        routeId: { type: string }
        fromLegId: { type: string }
        fromStopId: { type: string }
        toLegId: { type: string }
        toStopId: { type: string }
        ts: { type: string, format: date-time }
        changed: { type: boolean }
...